Make Hotel DMA slighly safer
The DMADescriptor still takes a raw pointer and size (through the flags
field), but at least the base DMA register only admits a `&'static
DMADescriptor` and the transfer size field is hidden (we don't use it in
scatter gather mode).

We ought to fix DMADescriptor but I think it'll affect how buffers need
to be managed, so leaving it as is for flexibility for now until we have
a couple more layers implemented.