Digest and AES now loading and running.
Digest passes the sha_test test.

AES does not pass the aes_test test. A change in the Driver trait in the
kernel means the system call interface for AES no longer works. It's
implemented strangely, with a callback for completion but then a
synchronous request to read the data. The synchronous request expects
a length to be returned, but that's no longer feasible (command only
returns ReturnCode).

I *think* the right way to do this is to give the buffer back (and
length, etc.) in the encrypt callback. This also removes the need for
these extra read syscalls.