This fixes a bug in the H1B UART driver. The observed behavior is that writes longer than the FIFO length can be truncated to the FIFO length. In the context of console, since there is an internal static buffer to copy userspace writes to, a long userspace printf appears as a series of truncated UART writes. E.g., if the FIFO is 32 bytes and the Console buffer is 64 bytes, then bytes 0-31, 64-95 will be actually written and 32-63 will be missed.
The bug occurs because the UART delivers an interrupt when an empty
FIFO is filled and the driver determines the write is complete when
it's not able to write any bytes to the FIFO; this completion
condition conflates there being no more bytes to send and there
being no space in the FIFO when an interrupt occurs.

The bugfix changes the return value of the underlying call such that
-1 means that the write completed, while 0 means no bytes were
written.

Note that this bug could be masked when there are multiple
outstanding writers, because writes might not encounter an
empty FIFO.