Correct the `offset` value in the flash transaction parameters.
The offsets are specified relative to the start of each flash module, whereas the h1b::hil::flash::Hardware interface uses offsets that cross over the two modules. H1bHw needs to do the mapping between them. Because only the second module is in use, we can simply hardcode the transformation.