Add NvCounter to golf2, and add an integration test.
I needed to tweak the flash API -- previously, Flash implementations were not guaranteed to return EBUSY when there was an outstanding programming request. The flash driver did have this behavior, and NvCounter relies on it, but FlashUser (part of the flash virtualization) did not. I documented that behavior in the Flash trait and modified FlashUser to implement busy check.