third_party/Build.mk: sandbox elf2tab build by copying source
The bubblewrap workaround for elf2tab's checked-in Cargo.lock file
messing up offline builds does not work in a docker container.  So a
more straightforward workaround is to simply copy the source tree for
elf2tab out of third_party/ and into build/ where we can simply remove
the Cargo.lock file before calling cargo build.  With this change, it
should be possible to build in a suitable docker container (that
contains the necessary tool chains) by invoking 'make BWRAP='.