Fix a race condition in building C apps.
The libtock-c application makefiles are not designed to be invoked concurrently for different apps. Doing so can result in linking errors, where one app's build system is re-building libraries (such as libh1) while another app's build system is trying to link against the library.

It does not appear that libtock-c's makefiles are designed to build multiple apps at once, so instead I decided to use a mutex to build apps one-at-a-time. I used flock for this, the same tool I used to serialize devicetests.

I tested this with several `make -j <all C apps>` invocations on my machine.