otpilot: Fix main loop to not use console before yieldk.
We used to print "yieldk()" to the console after checking conditions
but before actually calling yieldk() in the main loop. Since printing
to the console may result in calls to yieldk(), this resulted in the
main loop missing changed conditions and going into yieldk() again,
which could result in intermittent hangs.