Make builds under /tmp work.
Previously, the sandbox caused builds to fail (with a very confusing error message) if you checked out tock-on-titan under /tmp. The problem is the mount of /tmp in the sandbox hid the source code. With this change, /tmp and the source code are both visible and usable.

This change has no effect on builds outside /tmp.