kernel/h1: Add support for both flash banks.
This change adds support to the H1 flash driver for both flash banks.
In order to be able to use the existing interface / design, we now
need to pass the target address to both `H1bHw.set_transaction()` and
`H1bHw.trigger()`.

Further, this change adds `ReturnCode`s to the `Hardware` trait in
preparation to handle these in the driver code in a subsequent
commit.