kernel/h1: Add support for both flash banks. (#241)
This change adds support to the H1 flash driver for both flash banks. In order to be able to use the existing interface / design, we now need to pass the target bank to `H1bHw.trigger()` and pass the address relative to the target bank to `H1bHw.set_transaction()`. Test coverage for both banks has been added.